[move-prover] Aborts-with check mode.
This introduces a new property for `aborts_with` to check whether an aborts specification is complete:

```
aborts_with [check] code1, code2, ...
```

... checks whether the given function aborts with the given codes, independent of any other aborts specifications. That is, if there is already an `aborts_if P with code` it will not influence the outcome of the `aborts_with [check] codes` condition.

By default, the check can only verify whether there are any missing abort codes. In order to check whether they are obsolete codes which can never occur, a *negative* check needs to be performed by passing `--negative` to the prover's command line. Negative checks are based on a feature we added as part of Kevin's internship (we expect a verification error and report an error if it does *not* occur). Using the `--negative` flag is still experimental and may not work in all cases as expected.

Details:
- implements the `aborts_with [check]` functionality.
- introducing checks whether pragmas and properties are used correctly. Previously, a typo in a pragma or property would have been silently ignored; this is not longer the case.
- improvements on the negative verification infrastrucure.

Closes: #6121