[network] use a single `IntGauge` for per-peer channels
+ Previously, we were using a unique counter for each peer for each
channel that gets spun up for the `Peer` actor. This lead to a _ton_
of counters, even for very limited use cases like the validator network.
For example, there are ~10 channel counters per `Peer`. In premainnet,
each validator would be producing ~ 10 * 30 = 300 unique counters for
these metrics. When collected to the shared metrics, you get ~ 10 * 30 *
30 = 9000 counters.

+ This diff removes the peer_gauge in favor of a single counter for each
channel. Unfortunately this loses information (we can't isolate congestion
to a single channel x peer); however, it's probably fine...

+ Also clean up the channel docs and remove the shared global
TEST_COUNTER so the unit tests aren't as clunky.