[libra framework][docgen] Cleanup AccountFreezing and AccountLimits.
This PR cleans up documentation of mentioned modules. It also fixes specification incompleteness in AccountLimits and re-activates an invariant there. The incompleteness stemmed from turning some specs opaque but not determining all fields of a struct in post conditions.

Moreover, an issue in the doc generator with wrong indentation of `spec define` style helper functions is fixed. Also, we add a new option `references_file` to the generator which allows to add a file to each generated output containing reference definitions.

This is a first PR in a series; thought better split this down instead of one big PR to avoid merge conflicts.

Closes: #6314