[JSON-RPC] Adding a get_events_with_proof API
This is necessary to build good verifying clients and to ensure
a secure ecosystem upon launch.

Otherwise we are encouraging people to trust the full nodes.
This is adding a JSON-RPC call to get events with proofs, along
with the necessary backend changes to get these.

Notice that testing "get_events_with_proofs: start param is too big"
in the unit tests would require an implementation of the
get_events_with_proofs method in MockDB which does not exist yet.
Hence this test was omitted for now.

Closes: #6539