[storagte] state pruner: offload the first seek to the worker thread
PR #6803 avoided repeatedly seeking to the beginning in the stale node index, which was costly by doing it beforehand and only once at construction time. However that one time yet costly call sit in the critical path of the node start up, which can result in a slow node start if the pruning window was large.

This moves the initial seek to the work thread.

Closes: #7113