[type] switch to BCS bytes for txn args for the script function payload
NOTE: this change applies to the `ScriptFunction` payload type only, and
has no impact on how transaction arguments are encoded for the existing
`Script` payload type. We don't anticipate compatibility issues because
the `ScriptFunction` transaction types is not relased yet and hence, no
transactions can use this payload type.

One reason favoring BCS-encoding of txn args over using the existing
`TransactionArgument` enum for future development is extensibility:
when a new type of argument needs to be introduced, if we change the
`TransactionArgument` enum, this is a change to the whole system,
whereas a change to the generated boilerplate code and types is
generally much easier.

Transaction builders for all other languages (except TypeScript) are
updated as well, including: Java, Python, C++, C#, Go, and Rust.
TypeScript support is marked as TODO currently due to lack of test
cases.

Closes: #7973