[diem-swarm] Add proper support for VFN & PFN
Validator full nodes and Public full nodes had no difference in the
DiemSwarm framework.  This led to a bunch of confusion and skipping over
checking if they're connected.  Now, we properly check that there is at
least one outbound proper outbound connection for each full node, and
that the validator full nodes are connected to validators.