[smoke-test] Add initial framework for verifying client equiv test
The idea here is to whitebox fuzz the VerifyingClient and non-verifying
Client with the same random requests and assert that they both handle
the same requests identically.

Ideally, we should be able to just add new cases to the `arb_request`
strategy to test new request types as we add support for them.