[stackless-vm] adapt Move VM change_set for comparison
The original plan to have the stackless VM producing the same ChangeSet
as the Move VM is challenging. The reason can be evident in the
following program:

```
address 0x2::Test {
    use 0x1::Signer;

    struct S has key {
        val: u64,
    }

    fun mut_s_do_nothing(_s: &mut S) {
        // do nothing here
    }

    fun mut_s_same_value(s: &mut S) {
        // override with the same value
        s.val = s.val;
    }

    public(script) fun main1(account: signer) acquires S {
        let s = borrow_global_mut(Signer::address_of(&account));
        mut_s_do_nothing(s);
    }

    public(script) fun main2(account: signer) acquires S {
        let s = borrow_global_mut(Signer::address_of(&account));
        mut_s_same_value(s);
    }
}
```

The Move VM will not produce an entry in the ChangeSet if `main1` is
called but will produdce an entry if `main2` is called, although in both
cases, the underlying value of `S` is not changed.

However, in the view of the Stackless VM, both `main1` and `main2`
observes a write-back to the global state, so either a change set entry
is produced (in the view that the resource is modified) in both cases or
no entry should be produced in both cases.

We adopt the latter approach, i.e., no entries produced, in the view
that the underlying value stays the same. But this also means that we
need to adapt the changeset produced by the Move VM to match this
behavior in order for cross-vm comparison. This is what this commit is
about.