[stackless-vm] handle SaveMem in a separete EvalState that is dedicated for expression evaluation
The original plan is to record saved memory in GlobalState. However,
that will break the mental match that the GlobalState corresponds to the
Move global storage. As a result, we creaetd an EvalState to track and
record information produced in the expression evaluation (while the
GlobalState tracks information produced in bytecode interpretation)