[json-rpc] Avoid extra get_currencies lookup for get_account
Following the previous commit, we can avoid doing an extra
`get_currencies` call in `get_account` since we no longer need the
list of currencies to index into the current balance resources. The
same is true for the verifying client's implementation of `get_account`.
This halves the number of requests needed to support the verifying
client's `get_account` call.

Closes: #8415