[storage] Return consistency proof only for client-verifiable range
When a verifying client is syncing via get_state_proof requests, they
can't verify the accumulator consistency proof when there are more epoch
changes needed to sync before entering the epoch of that latest li. The
reason is that the current behaviour always returns a consistency proof
to the latest li version, regardless of whether or not the client will
actually be able to reach the latest li after applying the epoch
changes.

This diff fixes the above issue by limiting the consistency proof to the
last client-verifiable version. If the state proof manages to get the
client to the latest li, then the consistency proof is as before.
Otherwise, if the client will need more epoch changes, then the
consistency proof is only to the last epoch change li version.