[move-prover] a local monomorphization processor
We introduce the concept of *local monomorphization*, i.e., specializing
a generic proposition with types found in the enclosing function only.

Local monomorphization can only be performed when two conditions hold:

- all generic proposition are in the form of top-level unifersal type
  quantifiers, i.e., the proposition `P` must be in the form of
  `forall t1: type, t2: type, ... : expr<t1, t2, ...>` and there is no
  operation over P (e.g., !P, P ==> Q, etc is not allowed). This is
  inperfectly enforced in this commit (via a check that the kind of the
  quantifier to be eliminated must be a forall). A more robust solution
  is to use the generic invariant notion `invariant<T>`.`

- the global invariant instrumentation pass places a global invariant
  `I` into all relevant instantiations of a function `F`. For example,
  if `I<T>` talks about a memory `S<T>` that is also going to be touched
  by `F<T>`, then `I<T>` must be instrumented in the generic version of
  `F` as well as in every instantiation of `F` (e.g., `F<bool>`,
  `F<u64>`, etc).

If both conditions hold, we can run local monomorphization, i.e., assert
proposations that are only relevant to the enclosing function only and
safely assume that the proposation holds for all other types that are
not touched by the function.

This commits implements such a local monomorphization processor.