[move-model] add an option to ignore pragma opaque at ast level
`mvp --ignore-pragma-opaque-when-possible` will completely ignore any
`pragma opaque` specified in functions that
1) are not `public` or `public(script)` (i.e., all callers of the
   function can be found)
2) do not have a `[concrete]` or `[abstract]` property in its function spec

Here, "completely ignore" means that the move-model builder will treat
it as if the pragma does not exist in the ast level.

To support this, we added a new option type, `ModelBuilderOptions`, that
will be inserted to `GlobalEnv` before the model builder is invoked and
during the model building process, the options can be retrieved via
`GlobalEnv::get_extension<ModelBuilderOptions>()`.

This commit also replaces the `is_public` field in `FunEntry` with a
`visibility` field, as we now have more than just public/private
visibilities.