[move-prover] add a flag to toggle whether unconditional abort is inconsistency
The motivation for this change is the inconsistency issue reported in
PR #8851, where `TransactionFee::burn_fees<CoinType>` runs into an
unconditional abort case when and only when `CoinType` is instantiated
with `XDX`.

This is subsequently flagged by the inconsistency checker as an
violation. Where in this case, the inconsistency checker flags the
case not because we are not proving an `assert false;`, but because the
verification finishes successfully when the inconsistency checker
expects it to fail.

The reason verification finishes successfully is because for a function
that unconditionally abort on all arguments (and hence, never returns),
the `assert false;` verification condition is either not instrumented
or not reachable.

The fix, is to not only instrument `assert false;` before returns, but
also before an abort. In this way, we will be able to catch executions
that terminates without going through the return path.

However, we may still want to preserve the ability to check functions
that unconditionally abort, as we can prove function like the following
```
fun always_abort() {
    abort 0
}
spec always_abort {
    ensures 1 == 2;
}
```
does look a bit odd to users.

The compromise is to have an additional flag,
`--unconditional-abort-as-inconsistency`
to control whether we want to treat a function that aborts
unconditionally as an inconsistency violation. Unfortunately, due to the
complication in `TransactionFee.move`, we cannot turn on this flag by
default and have to rely on periodic manual runs to check these cases.

Closes: #8861