[move traits] Move VM MoveStorage trait to move-core-types
Increasingly, Move tools are relying on the MoveStorage trait for fetching modules and resources from on-chain. This trait lives in the VM, but tooling should not need to pull in the VM as a dependency just to use this trait.

This PR moves the trait to move-core-types and does some renaming to disambiguate from the MoveStorage trait in storage-interface. Additionally, it makes the error type generic instead of forcing implementers to return VM-internal errors.

One thing we might want to look at in the future is fixing the inconsistency between `get_module` (which takes a `ModuleId`) and `get_resource` (which takes an address and type). I think we should either change `get_module` to take an address and a name or change `get_resource` to take a `ResourceKey`).

Closes: #8886