[move-prover] record directly used memory and handle extra bytecode
There are two changes in this commit:

1/ Current implementation only records all (directly + transitively)
used memory, this commit additionally record the directly accessed
memory into a separate set, in a similar spirit to how the
modified / directly-modified sets are currently differentiated.
This information will be useful later when we calculate how to defer
invariants from a callee to its caller (by enlarging the set that
represents callers' directly-accessed memory).

2/ Several bytecode that has a memory accessing/modifying semantic
are not included in the analysis, namely `OpaqueCallBegin/End` and
`WriteBack` to a `GlobalRoot`.
  - Not covering `WriteBack` is hardly a bug because we would not
  expect a `WriteBack` to a `GlobalRoot` without a `BorrowGlobal`
  somewhere, and the memory usage/modification info should already
  be captured in the `BorrowGlobal` instruction. Nevertheless, it
  does not hurt to re-add the same info again.
  - Memory accesses in opaque functions are not captured previously
  which seems a bit weird. At least from the global invariant point
  of view, if a global invariant is applicable in an opaque function
  and that opaque function happens to defer the invariants to its
  caller, we should at least know which memories are accessed by the
  callee function in order to determine which invariants are applicable.