[move-prover] Implement spec (ghost) vars by mapping them to regular memory.
This re-activates and refines the specification variable implementation of the Move prover.

Spec variables are declared as previously, but with the option to add an initializer:

```
spec module {
   global spec_var: num = 1; // optional initializer
}
```

An update of a spec var can appear in a function spec block:

```
fun f() ...
spec f {
  ...
  update spec_var = spec_var + 1; // assignment; no old
}
```

In contrast to the older defunct implementation, this one maps spec vars to regular global "ghost" memory. That is for each `global x: T`, there is a generated `struct Ghost$x has key { v: T }`. Access to spec vars is right-away mapped to accessing this memory (at address `@0`) in move-model. This way, memory usage and global invariant analysis can be applied to spec vars uniformely.

For a number of examples, see the test [global_vars.move]().

This implementation isn't complete yet; there is some more work to do with opaque functions.

Closes: #9154