[move-prover] dump bytecode and result in output dir
Previous dumping location is at the first Move source location, which
may pollute the source directories. Furthermore, if we pass a directory
as the first source location, the process will panic.

This commit changes the default output location for `--dump-bytecode`
to the parent directory of the `output.bpl` file, and format ihe
bytecode dumps with `bytecode_{step_number}_{step_name}.bytecode`.

Closes: #9052