[move-stdlib] Moving Capability module to stdlib and refining it
This moves `Capability` out of nursery into stdlib so we can use it in experimentation with new access control patterns.

It also refines the implementation and documentation:

- Now every public function is protected by a feature witness. The using module is expected to build wrappers as required for the specific scenario. The documentation example has been updated to reflect this.
- `LinearCap` has been introduced. This is just an alternative wrapper, whether linear or copying capabilities are exposed is decided by the module which owns the feature.
- Delegation and revocation functions have been made idempotent.
- Unit tests have been added.

I'm not yet sure how to deal with the problem of delegating to the same address multiple capabilities of the same type. One solution on the table was suggested by Junkil, having essentially `acquires_with_known_root(signer, root)`, but I think this need further iteration.

Closes: #9328