[base_image] One-line fix in scripts/dev_setup.sh
./scripts/dev_setup.sh -yp installs Move Prover tools such as Boogie,
and sets up a .profile on the users directory to set shell variables
so that the Prover can call them.

For me, the prover tool were being installed in one place, and the
.profile was pointing to another place.

The problem was that OPT_DIR is assigned "true" or "false", and the
test was -n "$OPT_DIR" instead of "$OPT_DIR" == "true".  I changed it.

Closes: #9433