[move] Language server initialization
Implement language server initialization and an event loop in the
move-analyzer language server executable. In tandem, implement a client
for the server in the move-analyzer VS Code extension.

For now, the server only advertises a single capability: "go to
definition." This capability causes VS Code to display a "go to
definition" command option when users have the move-analyzer extension
installed. However, for now, when that option is used, the server panics
(implementing a response is a "TODO").

However, despite the panic, several "features" are added as part of this
commit: users of the VS Code extension will see logs in their "Move
Analyzer Client" and "Move Language Server" output channels indicating
that the client and server have initialized and connected to one
another.

Closes: #9463